+++ Source/dsp/StateMachine.h
@@ -1,0 +1,15 @@
+#pragma once
+
+enum class LoopState
+{
+    Idle,
+    Record,
+    Play,
+    Overdub
+};
+
+class StateMachine
+{
+public:
+    LoopState update(bool record, bool play, bool overdub, bool hasLoop, bool clearEdge);
+};